Criação do BoraMorar.Moradias

+++ BoraMorar/Imoveis/Repository/IImovelRepository.cs
@@ -2,5 +2,5 @@
 
 public interface IImovelRepository : IAddRepository<Imovel>
 {
-    Task<Imovel?> FindAsync(int id);
+    Task<Imovel?> FindAsync(long id);
 }
